Even more better code

+++ kernel/include/gfx/framebuffer.h
@@ -1,8 +1,9 @@
 #pragma once
 #include <boot/limine.h>
+#include <stddef.h>
 #include <stdint.h>
 
-void putpixel(int x, int y, uint32_t c);
-void putchar16(int x, int y, char ch, uint32_t color);
-void putstr16(int x, int y, char *s, uint32_t color);
+void putpixel(size_t x, size_t y, uint32_t c);
+void putchar16(size_t x, size_t y, char ch, uint32_t color);
+void putstr16(size_t x, size_t y, char *s, uint32_t color);
 void init_fb(struct limine_framebuffer *);
